인증 모듈 분리

+++ controller/LikeController.js
@@ -1,46 +1,60 @@
+const ensureAuthorization = require("../auth"); // 인증 모듈
 const jwt = require("jsonwebtoken");
-const conn = require("../mariadb");
-const { StatusCodes } = require("http-status-codes");
+const conn = require("../mariadb"); // db 모듈
+const { StatusCodes } = require("http-status-codes"); // status code 모듈
 const dotenv = require("dotenv");
 dotenv.config();
 
 const addLike = (req, res) => {
   const book_id = req.params.id;
-  const authorization = ensureAuthorization(req);
+  const authorization = ensureAuthorization(req, res);
 
-  const sql = `INSERT INTO likes(user_id, liked_book_id) VALUES (?, ?);`;
-  const values = [authorization.id, book_id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res
+      .status(StatusCodes.UNAUTHORIZED)
+      .json({ message: "로그인 세션이 만료되었습니다. 다시 로그인 해주세요." });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res
+      .status(StatusCodes.BAD_REQUEST)
+      .json({ message: "잘못된 토큰입니다." });
+  } else {
+    const sql = `INSERT INTO likes(user_id, liked_book_id) VALUES (?, ?);`;
+    const values = [authorization.id, book_id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const removeLike = (req, res) => {
   const book_id = req.params.id;
-  const authorization = ensureAuthorization(req);
+  const authorization = ensureAuthorization(req, res);
 
-  const sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;`;
-  const values = [authorization.id, book_id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res
+      .status(StatusCodes.UNAUTHORIZED)
+      .json({ message: "로그인 세션이 만료되었습니다. 다시 로그인 해주세요." });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res
+      .status(StatusCodes.BAD_REQUEST)
+      .json({ message: "잘못된 토큰입니다." });
+  } else {
+    const sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?;`;
+    const values = [authorization.id, book_id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
-
-function ensureAuthorization(req) {
-  const receivedJwt = req.headers["authorization"];
-  const decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-
-  return decodedJwt;
-}
 
 module.exports = { addLike, removeLike };
